A slightly naive approach to making things faster.
We can just only call  when we're editing a new variable and things go a lot faster.

But there's a problem. Play with the sliders for a while and you'll find you get into a weird state. If you change them in certain orders in certain ways, they get stuck. All of a sudden total won't let you drag it below the current value of dogs, or cats can't be increased above 5. The specifics differ based on what you change in what order, but eventually the system gets stuck in one spot rather than adjusting naturally to your input.